Add events database caching feature
Implemented a feature to cache events received from the Janus server into a database. Provided a MySQL configuration as an example, with the possibility of adding other database types easily. Also included information on the types of events that can be cached and how to access them in the updated ReadMe file.