Update app version and add method descriptions to Javadoc
This commit updates the SDK version from 0.5 to 0.6 in various Javadoc HTML files and the build.gradle file. It also adds or updates descriptions for several methods within the Javadoc and test files. In addition to these changes, new methods related to PostProcessing interface are introduced in the Javadoc index file.